model_preparation/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
click>=8.1.0
pytest>=7.4.0
streamlit>=1.28.0

